requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.2.0
black==20.8b1
chardet==3.0.4
click==7.1.2
discord.py==1.5.0
idna==2.10
multidict==4.7.6
mypy-extensions==0.4.3
pathspec==0.8.0
pkg-resources==0.0.0
regex==2020.10.15
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
yarl==1.5.1
